+++ log.go
@@ -8,46 +8,6 @@ import (
 	context "golang.org/x/net/context"
 )
 
-// StandardLogger provides API compatibility with standard printf loggers
-// eg. go-logging
-type StandardLogger interface {
-	Debug(args ...interface{})
-	Debugf(format string, args ...interface{})
-	Error(args ...interface{})
-	Errorf(format string, args ...interface{})
-	Fatal(args ...interface{})
-	Fatalf(format string, args ...interface{})
-	Info(args ...interface{})
-	Infof(format string, args ...interface{})
-	Panic(args ...interface{})
-	Panicf(format string, args ...interface{})
-	Warning(args ...interface{})
-	Warningf(format string, args ...interface{})
-}
-
-// EventLogger extends the StandardLogger interface to allow for log items
-// containing structured metadata
-type EventLogger interface {
-	StandardLogger
-
-	// Event merges structured data from the provided inputs into a single
-	// machine-readable log event.
-	//
-	// If the context contains metadata, a copy of this is used as the base
-	// metadata accumulator.
-	//
-	// If one or more loggable objects are provided, these are deep-merged into base blob.
-	//
-	// Next, the event name is added to the blob under the key "event". If
-	// the key "event" already exists, it will be over-written.
-	//
-	// Finally the timestamp and package name are added to the accumulator and
-	// the metadata is logged.
-	Event(ctx context.Context, event string, m ...Loggable)
-
-	EventBegin(ctx context.Context, event string, m ...Loggable) *EventInProgress
-}
-
 // Logger retrieves an event logger by name
 func Logger(system string) EventLogger {
